python_server/requirements.txt

websockets==12.0
psutil==5.9.8
python-dotenv==1.0.1
